fix: #22 windows os download rename err

+++ api/weapi/comment.go
@@ -1,3 +1,26 @@
+// MIT License
+//
+// Copyright (c) 2025 chaunsin
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+
 package weapi
 
 import (
@@ -20,14 +43,14 @@ type CommentInfoListRespData struct {
 	LatestLikedUsers  interface{} `json:"latestLikedUsers"`
 	Liked             bool        `json:"liked"`
 	Comments          interface{} `json:"comments"`
-	ResourceType      int         `json:"resourceType"`
-	ResourceId        int         `json:"resourceId"` // 应改是对应的歌曲id
+	ResourceType      int64       `json:"resourceType"`
+	ResourceId        int64       `json:"resourceId"` // 应改是对应的歌曲id
 	CommentUpgraded   bool        `json:"commentUpgraded"`
-	MusicianSaidCount int         `json:"musicianSaidCount"`
+	MusicianSaidCount int64       `json:"musicianSaidCount"`
 	CommentCountDesc  string      `json:"commentCountDesc"` // 评论数描述基本和commentCount一样
-	LikedCount        int         `json:"likedCount"`
-	CommentCount      int         `json:"commentCount"` // 评论数
-	ShareCount        int         `json:"shareCount"`
+	LikedCount        int64       `json:"likedCount"`
+	CommentCount      int64       `json:"commentCount"` // 评论数
+	ShareCount        int64       `json:"shareCount"`
 	ThreadId          string      `json:"threadId"` // 线程id，用于获取评论列表使用
 }
 
@@ -70,53 +93,53 @@ type CommentsReq struct {
 
 type CommentsResp struct {
 	IsMusician  bool          `json:"isMusician"`
-	Cnum        int           `json:"cnum"`
-	UserId      int           `json:"userId"`
+	Cnum        int64         `json:"cnum"`
+	UserId      int64         `json:"userId"`
 	TopComments []interface{} `json:"topComments"`
 	Code        int64         `json:"code"`
 	Comments    []struct {
 		User struct {
 			LocationInfo interface{} `json:"locationInfo"`
 			LiveInfo     interface{} `json:"liveInfo"`
-			Anonym       int         `json:"anonym"`
+			Anonym       int64       `json:"anonym"`
 			Highlight    bool        `json:"highlight"`
 			AvatarUrl    string      `json:"avatarUrl"`
 			AvatarDetail *struct {
-				UserType        int    `json:"userType"`
-				IdentityLevel   int    `json:"identityLevel"`
+				UserType        int64  `json:"userType"`
+				IdentityLevel   int64  `json:"identityLevel"`
 				IdentityIconUrl string `json:"identityIconUrl"`
 			} `json:"avatarDetail"`
-			UserType     int         `json:"userType"`
+			UserType     int64       `json:"userType"`
 			Followed     bool        `json:"followed"`
 			Mutual       bool        `json:"mutual"`
 			RemarkName   interface{} `json:"remarkName"`
 			SocialUserId interface{} `json:"socialUserId"`
 			VipRights    *struct {
 				Associator *struct {
-					VipCode int    `json:"vipCode"`
+					VipCode int64  `json:"vipCode"`
 					Rights  bool   `json:"rights"`
 					IconUrl string `json:"iconUrl"`
 				} `json:"associator"`
 				MusicPackage *struct {
-					VipCode int    `json:"vipCode"`
+					VipCode int64  `json:"vipCode"`
 					Rights  bool   `json:"rights"`
 					IconUrl string `json:"iconUrl"`
 				} `json:"musicPackage"`
 				Redplus *struct {
-					VipCode int    `json:"vipCode"`
+					VipCode int64  `json:"vipCode"`
 					Rights  bool   `json:"rights"`
 					IconUrl string `json:"iconUrl"`
 				} `json:"redplus"`
-				RedVipAnnualCount int         `json:"redVipAnnualCount"`
-				RedVipLevel       int         `json:"redVipLevel"`
-				RelationType      int         `json:"relationType"`
+				RedVipAnnualCount int64       `json:"redVipAnnualCount"`
+				RedVipLevel       int64       `json:"redVipLevel"`
+				RelationType      int64       `json:"relationType"`
 				MemberLogo        interface{} `json:"memberLogo"`
 			} `json:"vipRights"`
 			Nickname       string      `json:"nickname"`
-			AuthStatus     int         `json:"authStatus"`
+			AuthStatus     int64       `json:"authStatus"`
 			ExpertTags     interface{} `json:"expertTags"`
 			Experts        interface{} `json:"experts"`
-			VipType        int         `json:"vipType"`
+			VipType        int64       `json:"vipType"`
 			CommonIdentity interface{} `json:"commonIdentity"`
 			UserId         int64       `json:"userId"`
 			Target         interface{} `json:"target"`
@@ -125,25 +148,25 @@ type CommentsResp struct {
 			User struct {
 				LocationInfo interface{} `json:"locationInfo"`
 				LiveInfo     interface{} `json:"liveInfo"`
-				Anonym       int         `json:"anonym"`
+				Anonym       int64       `json:"anonym"`
 				Highlight    bool        `json:"highlight"`
 				AvatarUrl    string      `json:"avatarUrl"`
 				AvatarDetail *struct {
-					UserType        int    `json:"userType"`
-					IdentityLevel   int    `json:"identityLevel"`
+					UserType        int64  `json:"userType"`
+					IdentityLevel   int64  `json:"identityLevel"`
 					IdentityIconUrl string `json:"identityIconUrl"`
 				} `json:"avatarDetail"`
-				UserType       int         `json:"userType"`
+				UserType       int64       `json:"userType"`
 				Followed       bool        `json:"followed"`
 				Mutual         bool        `json:"mutual"`
 				RemarkName     interface{} `json:"remarkName"`
 				SocialUserId   interface{} `json:"socialUserId"`
 				VipRights      interface{} `json:"vipRights"`
 				Nickname       string      `json:"nickname"`
-				AuthStatus     int         `json:"authStatus"`
+				AuthStatus     int64       `json:"authStatus"`
 				ExpertTags     interface{} `json:"expertTags"`
 				Experts        interface{} `json:"experts"`
-				VipType        int         `json:"vipType"`
+				VipType        int64       `json:"vipType"`
 				CommonIdentity interface{} `json:"commonIdentity"`
 				UserId         int64       `json:"userId"`
 				Target         interface{} `json:"target"`
@@ -151,7 +174,7 @@ type CommentsResp struct {
 			BeRepliedCommentId int64       `json:"beRepliedCommentId"`
 			Content            *string     `json:"content"`
 			RichContent        *string     `json:"richContent"`
-			Status             int         `json:"status"`
+			Status             int64       `json:"status"`
 			ExpressionUrl      interface{} `json:"expressionUrl"`
 			IpLocation         struct {
 				Ip       interface{} `json:"ip"`
@@ -160,11 +183,11 @@ type CommentsResp struct {
 			} `json:"ipLocation"`
 		} `json:"beReplied"`
 		PendantData *struct {
-			Id       int    `json:"id"`
+			Id       int64  `json:"id"`
 			ImageUrl string `json:"imageUrl"`
 		} `json:"pendantData"`
 		ShowFloorComment    interface{} `json:"showFloorComment"`
-		Status              int         `json:"status"`
+		Status              int64       `json:"status"`
 		CommentId           int64       `json:"commentId"`
 		Content             string      `json:"content"` // 评论内容
 		RichContent         *string     `json:"richContent"`
@@ -172,9 +195,9 @@ type CommentsResp struct {
 		Time                int64       `json:"time"`
 		TimeStr             string      `json:"timeStr"`
 		NeedDisplayTime     bool        `json:"needDisplayTime"`
-		LikedCount          int         `json:"likedCount"`
+		LikedCount          int64       `json:"likedCount"`
 		ExpressionUrl       interface{} `json:"expressionUrl"`
-		CommentLocationType int         `json:"commentLocationType"`
+		CommentLocationType int64       `json:"commentLocationType"`
 		ParentCommentId     int64       `json:"parentCommentId"`
 		Decoration          struct {
 		} `json:"decoration"`
@@ -192,8 +215,8 @@ type CommentsResp struct {
 		} `json:"likeAnimationMap"`
 		Liked bool `json:"liked"`
 	} `json:"comments"`
-	Total int  `json:"total"`
-	More  bool `json:"more"`
+	Total int64 `json:"total"`
+	More  bool  `json:"more"`
 }
 
 // Comments 获取歌曲评论列表
